cleanup(ε.2): extract decode_block/spectrogram.rs (Spectrogram + compute_spectrogram)
Second slice of the `decode_block.rs` per-stage split. Moves the
spectrogram container + both backends into a sibling submodule:

- `Spectrogram` struct + `from_parts()`
- `SpecCell` (f32 / u16 cfg-gated) type alias
- `FP_SPEC_SHIFT` constant (private to the module — only the
  fixed-point compute_spectrogram body references it)
- `CoarseAcc` type alias + `Spectrogram::power_acc()` method
- `compute_spectrogram<S>()` host (fft-rustfft) backend
- `compute_spectrogram<S>()` embedded fixed-point backend (incl. the
  two-for-one real-FFT trick + Hann pre-window + auto-gain shift)

External-caller path preserved: parent re-exports `Spectrogram`,
`SpecCell`, `compute_spectrogram` via `pub use spectrogram::*`.
`CoarseAcc` + `power_acc` are surfaced as `pub(super)` for the
coarse-sync code that still lives in the parent — they'll move with
ε.3 (coarse_sync extraction).

NMAX import dropped from the parent's top-level `use super::params::*`
and re-imported inside the `#[cfg(test)] mod tests` (its only
remaining consumer post-split).

Acceptance:
- `cargo check` clean on all 4 feature cells
- FT8 qso3_busy golden tests pass (WSJT-X / JTDX / AP-on diag)
- 7 `ft8::decode_block::tests::*` lib tests pass
- `cargo clippy --features full --lib` clean, `cargo fmt --check` clean
- Diff: decode_block.rs −338 / +8, new decode_block/spectrogram.rs +357
  (mostly verbatim copy plus the imports the moved code needs)

Stacks on #77 (ε.1). Will rebase to main when #77 lands.
